+++ android/app/src/main/java/org/findmyfam/ui/groups/GroupChatScreen.kt
@@ -60,9 +60,7 @@ fun GroupChatScreen(
             TopAppBar(
                 title = {
                     Column(
-                        modifier = Modifier
-                            .fillMaxHeight()
-                            .clickable(onClick = onDetail)
+                        modifier = Modifier.clickable(onClick = onDetail)
                     ) {
                         Text(
                             text = groupName,